+++ src/callbacks/postNewRoom.ts
@@ -18,6 +18,7 @@ import { seededDeathPostNewRoom } from "../features/mandatory/seededDeath/callba
 import * as tempMoreOptions from "../features/mandatory/tempMoreOptions";
 import * as trophy from "../features/mandatory/trophy";
 import * as fastSatan from "../features/optional/bosses/fastSatan";
+import * as preventUltraSecretRoomSoftlock from "../features/optional/bugfix/preventUltraSecretRoomSoftlock";
 import * as removeInvalidPitfalls from "../features/optional/bugfix/removeInvalidPitfalls";
 import * as teleportInvalidEntrance from "../features/optional/bugfix/teleportInvalidEntrance";
 import * as showEdenStartingItems from "../features/optional/characters/showEdenStartingItems";
@@ -98,6 +99,7 @@ export function main(): void {
   extraStartingItemsPostNewRoom();
 
   // Bug fixes
+  preventUltraSecretRoomSoftlock.postNewRoom();
   teleportInvalidEntrance.postNewRoom();
   removeInvalidPitfalls.postNewRoom();