+++ src/types/ActiveCollectibleDescription.ts
@@ -1,4 +1,4 @@
-export interface ActiveItemDescription {
+export interface ActiveCollectibleDescription {
   collectibleType: CollectibleType | int;
   charge: int;
   batteryCharge: int;